rename copied FieldHeading to LabeledField
